+++ src/string.rs
@@ -1,25 +1,10 @@
+use crate::generic::RedisGeneric;
 use crate::traits::BackedType;
 use redis::{Commands, RedisResult};
 use std::ops;
 use std::ops::AddAssign;
 
-pub struct TString {
-    value: String,
-    field_name: String,
-    pub(crate) conn: Option<redis::Connection>,
-    pub(crate) client: redis::Client,
-}
-
-impl TString {
-    pub fn new(value: String, client: redis::Client) -> TString {
-        TString {
-            value,
-            client,
-            conn: None,
-            field_name: uuid::Uuid::new_v4().to_string(),
-        }
-    }
-}
+pub type TString = RedisGeneric<String>;
 
 impl ops::Add<TString> for TString {
     type Output = TString;
@@ -46,6 +31,12 @@ impl AddAssign<TString> for TString {
     }
 }
 
+impl PartialEq<&str> for TString {
+    fn eq(&self, other: &&str) -> bool {
+        self.value == *other
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -53,14 +44,22 @@ mod tests {
     #[test]
     fn test_string() {
         let client = redis::Client::open("redis://localhost/").unwrap();
-        let mut s1 = TString::new("Hello".to_string(), client.clone());
-        let mut s2 = TString::new("World".to_string(), client.clone());
-        let s3 = TString::new("Together".to_string(), client.clone());
+        let mut s1 = TString::new("Hello".to_string(), client.clone(), "s1".to_string());
+        let mut s2 = TString::new("World".to_string(), client.clone(), "s2".to_string());
+        let s3 = TString::new("Together".to_string(), client.clone(), "s3".to_string());
         assert_eq!(s1.value, "Hello");
         assert_eq!(s2.value, "World");
         s1 += s2;
         assert_eq!(s1.value, "HelloWorld");
         s2 = s1 + s3;
         assert_eq!(s2.value, "HelloWorldTogether");
+    }
+
+    #[test]
+    fn test_partialeq() {
+        let client = redis::Client::open("redis://localhost/").unwrap();
+        let mut s1 = TString::new("Hello".to_string(), client.clone(), "s1".to_string());
+        assert_eq!(s1, "Hello");
+        assert_ne!(s1, "World");
     }
 }